Add files via upload
this version introduces the removal of words that are using multiple of the same letter with the assumption that you only have one of each of your letters. This assumption is obviously flawed but I went with it for now and the ability to truly tell what a player is working with is to come, e.g. a players has an 'a', 'l', 's', 'o'. That player can build the word also but not alas, however with the code as written it will remove alas from the possible words even if the user enters 'a', 'a', 'l', 's'. 

It also now scores the words remaining in the list of playable words and will tell you your best options.

It also now